+++ packages/chakra-ui-core/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 0.6.1
+
+### Patch Changes
+
+- ## Bug fixes
+  - `CCollapse` component name option typo fix
+  - `CMenu`: add `as` polymophic prop support for `CMenuItem` component
+  - `chakra-loader`: Add documentation for the `chakra-loader` webpack plugin
+  - Community: Add link to Chakra UI discord to docs.
+
 ## 0.6.0
 
 ### Minor Changes